Update and rename testing.py to testing2.py
This testing is the version 2 with new test cases
